+++ internal/server/handlers/common.go
@@ -4,6 +4,7 @@ import (
 	er "github.com/TanmoySG/wunderDB/internal/errors"
 	"github.com/TanmoySG/wunderDB/internal/users/authentication"
 	"github.com/TanmoySG/wunderDB/model"
+	"github.com/go-playground/validator/v10"
 	"github.com/gofiber/fiber/v2"
 )
 
@@ -62,4 +63,20 @@ func (wh wdbHandlers) handleAuthorization(username string, entity model.Entities
 	}
 
 	return authSuccessful, nil
+}
+
+func SendResponse(c *fiber.Ctx, marshaledResponse []byte, statusCode int) error {
+	c.Set(ContentType, ApplicationJson)
+	c.Send(marshaledResponse)
+	return c.SendStatus(statusCode)
+}
+
+func ValidateRequest(request any) *er.WdbError {
+	validate := validator.New()
+
+	err := validate.Struct(request)
+	if err != nil {
+		return &er.ValidationError
+	}
+	return nil
 }
