+++ internal/server/handlers/handlers.go
@@ -1,12 +1,7 @@
 package handlers
 
 import (
-	"fmt"
-
-	er "github.com/TanmoySG/wunderDB/internal/errors"
-	"github.com/TanmoySG/wunderDB/internal/server/response"
 	w "github.com/TanmoySG/wunderDB/pkg/wdb"
-	"github.com/go-playground/validator/v10"
 	"github.com/gofiber/fiber/v2"
 )
 
@@ -48,28 +43,4 @@ func NewHandlers(client w.Client) Client {
 	return wdbHandlers{
 		wdbClient: client,
 	}
-}
-
-func (wh wdbHandlers) Hello(c *fiber.Ctx) error {
-	msg := map[string]string{
-		"message": fmt.Sprintf("✋ %s", "hello"),
-	}
-	resp := response.Format("ping", nil, msg)
-	return SendResponse(c, resp.Marshal(), resp.HttpStatusCode)
-}
-
-func SendResponse(c *fiber.Ctx, marshaledResponse []byte, statusCode int) error {
-	c.Set(ContentType, ApplicationJson)
-	c.Send(marshaledResponse)
-	return c.SendStatus(statusCode)
-}
-
-func ValidateRequest(request any) *er.WdbError {
-	validate := validator.New()
-
-	err := validate.Struct(request)
-	if err != nil {
-		return &er.ValidationError
-	}
-	return nil
 }
